commented hw06 code

+++ homework/hw6/hw06-single.cpp
@@ -13,12 +13,16 @@ using namespace std;
 
 class Noble; // forward declaration
 
+/**
+ * @brief The Warrior class
+ *
+ */
 class Warrior {
 public:
-    Warrior(const string& name, int strength);
+    Warrior(const string& name, double strength);
     const string& getName() const;
-    const int getStrength() const;
-    void setStrength(int strength);
+    const double getStrength() const;
+    void setStrength(double strength);
     const Noble* getEmployer() const;
     void setEmployer(Noble* employer);
     bool isHired() const;
@@ -26,10 +30,14 @@ public:
     friend ostream& operator<<(ostream& os, const Warrior& warrior);
 private:
     string name;
-    int strength;
+    double strength;
     Noble* employer;
 };
 
+/**
+ * @brief The Noble class
+ *
+ */
 class Noble {
 public:
     Noble(const string& name);
@@ -46,6 +54,11 @@ private:
     vector<Warrior*> army;
 };
 
+/**
+ * @brief Simulates a game of medival times, as per the problem statement
+ *
+ * @return int 0 upon success
+ */
 int main() {
     Noble art("King Arthur");
     Noble lance("Lancelot du Lac");
@@ -106,33 +119,76 @@ int main() {
     cout << "===============================================\n\n";
 }
 
-Warrior::Warrior(const string& name, int strength):
+/**
+ * @brief Construct a new Warrior object
+ *
+ * @param name the name of the warrior
+ * @param strength the strength of the warrior
+ */
+Warrior::Warrior(const string& name, double strength):
     name(name), strength(strength), employer(nullptr) {}
 
+/**
+ * @brief Get the name of the warrior
+ *
+ * @return const string& the name of the warrior
+ */
 const string& Warrior::getName() const {
     return name;
 }
 
-const int Warrior::getStrength() const {
+/**
+ * @brief Get the strength of the warrior
+ *
+ * @return const double the strength of the warrior
+ */
+const double Warrior::getStrength() const {
     return strength;
 }
 
-void Warrior::setStrength(int strength) {
+/**
+ * @brief Set the strength of the warrior
+ *
+ * @param strength the new strength of the warrior
+ */
+void Warrior::setStrength(double strength) {
     this->strength = strength;
 }
 
+/**
+ * @brief Get the employer of the warrior
+ *
+ * @return const Noble* the employer of the warrior
+ */
 const Noble* Warrior::getEmployer() const {
     return employer;
 }
 
+/**
+ * @brief Set the employer of the warrior
+ *
+ * @param employer the new employer of the warrior
+ */
 void Warrior::setEmployer(Noble* employer) {
     this->employer = employer;
 }
 
+/**
+ * @brief Check if the warrior is hired
+ *
+ * @return true if the warrior is hired
+ *         false otherwise
+ */
 bool Warrior::isHired() const {
     return employer != nullptr;
 }
 
+/**
+ * @brief Make the warrior runaway
+ *
+ * @return true if the warrior was successfully ran away,
+ *         false otherwise
+ */
 bool Warrior::runaway() {
     if (employer != nullptr) {
         cout << name << " flees in terror, abandoning his lord, "
@@ -146,17 +202,41 @@ bool Warrior::runaway() {
     }
 }
 
+/**
+ * @brief Overload the << operator for Warrior
+ *
+ * @param os the output stream
+ * @param warrior the warrior to be output
+ * @return ostream& the output stream
+ */
 ostream& operator<<(ostream& os, const Warrior& warrior) {
     os << warrior.name << ": " << warrior.strength;
     return os;
 }
 
+/**
+ * @brief Construct a new  Noble object
+ *
+ * @param name the name of the noble
+ */
 Noble::Noble(const string& name): name(name), strength(0), alive(true) {}
 
+/**
+ * @brief Get the name of the noble
+ *
+ * @return const string& the name of the noble
+ */
 const string& Noble::getName() const {
     return name;
 }
 
+/**
+ * @brief Hire a warrior
+ *
+ * @param warrior the warrior to be hired
+ * @return true if the warrior was successfully hired,
+ *         false otherwise
+ */
 bool Noble::hire(Warrior& warrior) {
     if (!alive) {
         cout << name << " is dead -- they cannot hire anymore." << endl;
@@ -174,6 +254,15 @@ bool Noble::hire(Warrior& warrior) {
     }
 }
 
+/**
+ * @brief Fire a warrior
+ *
+ * @param warrior the warrior to be fired
+ * @param runaway true if the warrior is 'fired' is running away,
+ *                false otherwise
+ * @return true if the warrior was successfully fired,
+ *         false otherwise
+ */
 bool Noble::fire(Warrior& warrior, bool runaway) {
     size_t warriorPosition = army.size();
     for (size_t i = 0; i < army.size(); ++i) {
@@ -202,6 +291,11 @@ bool Noble::fire(Warrior& warrior, bool runaway) {
     }
 }
 
+/**
+ * @brief Set the strength of all the noble's warriors to 0,
+ *       the noble's strength to 0, and the noble's alive status to false
+ * 
+ */
 void Noble::uponDeath() {
     for (size_t i = 0; i < army.size(); ++i) {
         army[i]->setStrength(0);
@@ -210,6 +304,11 @@ void Noble::uponDeath() {
     alive = false;
 }
 
+/**
+ * @brief Do battle with another noble
+ * 
+ * @param otherNoble the other noble to do battle with
+ */
 void Noble::battle(Noble& otherNoble) {
     cout << name << " battles " << otherNoble.name << endl;
     if (!alive && !otherNoble.alive) {
@@ -247,6 +346,13 @@ void Noble::battle(Noble& otherNoble) {
     }
 }
 
+/**
+ * @brief Overload the << operator for Noble
+ * 
+ * @param os the output stream
+ * @param noble the Noble to be output
+ * @return ostream& the output stream
+ */
 ostream& operator<<(ostream& os, const Noble& noble) {
     os << noble.name << " has an army of " << noble.army.size() << endl;
     for (size_t i = 0; i < noble.army.size(); ++i) {
